+++ src/Rdd.Infra/Storage/IPropertyAuthorizer.cs
@@ -1,0 +1,11 @@
+﻿using Rdd.Domain.Helpers.Expressions;
+
+namespace Rdd.Infra.Storage
+{
+    public interface IPropertyAuthorizer<TEntity>
+    {
+        IExpressionTree IncludeWhiteList { get; }
+
+        bool IsVisible(IExpressionChain property);
+    }
+}