+++ src/Rdd.Web/Helpers/RddServiceCollectionExtensions.cs
@@ -67,7 +67,8 @@ namespace Rdd.Web.Helpers
             services.TryAddSingleton<IExpressionParser, ExpressionParser>();
             services.TryAddSingleton(typeof(IWebFilterConverter<>), typeof(WebFilterConverter<>));
             services.TryAddSingleton<IPagingParser, PagingParser>();
-            services.TryAddSingleton<IFilterParser, FilterParser>();
+            services.TryAddSingleton(typeof(IPropertyAuthorizer<>), typeof(PropertyAuthorizer<>));
+            services.TryAddSingleton(typeof(IFilterParser<>), typeof(FilterParser<>));
             services.TryAddSingleton<IFieldsParser, FieldsParser>();
             services.TryAddSingleton<IOrderByParser, OrderByParser>();
             services.TryAddSingleton(typeof(IQueryParser<>), typeof(QueryParser<>));