+++ src/Rdd.Web/Querying/FilterParser.cs
@@ -9,6 +9,7 @@ using Rdd.Domain.Helpers.Expressions;
 using Rdd.Domain.Models;
 using Rdd.Domain.Models.Querying;
 using Rdd.Infra.Helpers;
+using Rdd.Infra.Storage;
 using Rdd.Infra.Web.Models;
 using Rdd.Web.Helpers;
 using System;
@@ -18,7 +19,8 @@ using System.Linq;
 
 namespace Rdd.Web.Querying
 {
-    public class FilterParser : IFilterParser
+    public class FilterParser<TEntity> : IFilterParser<TEntity>
+        where TEntity : class
     {
         protected static readonly IReadOnlyDictionary<StringSegment, WebFilterOperand> Operands = new Dictionary<StringSegment, WebFilterOperand>(StringSegmentComparer.OrdinalIgnoreCase)
         {
@@ -37,30 +39,35 @@ namespace Rdd.Web.Querying
 
         private readonly IStringConverter _stringConverter;
         private readonly IExpressionParser _expressionParser;
+        private readonly IWebFilterConverter<TEntity> _webFilterConverter;
+        private readonly IPropertyAuthorizer<TEntity> _propertyAuthorizer;
 
-        public FilterParser(IStringConverter stringConverter, IExpressionParser expressionParser)
+        public FilterParser(IStringConverter stringConverter, IExpressionParser expressionParser, IWebFilterConverter<TEntity> webFilterConverter, IPropertyAuthorizer<TEntity> propertyAuthorizer)
         {
             _stringConverter = stringConverter ?? throw new ArgumentNullException(nameof(stringConverter));
             _expressionParser = expressionParser ?? throw new ArgumentNullException(nameof(expressionParser));
+            _webFilterConverter = webFilterConverter ?? throw new ArgumentNullException(nameof(webFilterConverter));
+            _propertyAuthorizer = propertyAuthorizer ?? throw new ArgumentNullException(nameof(propertyAuthorizer));
         }
 
-        public virtual WebFilter<TEntity> Parse<TEntity>(string key, string value)
+        public virtual WebFilter<TEntity> Parse(string key, string value)
         {
+            var chain = _expressionParser.ParseChain<TEntity>(key);
+            if (!_propertyAuthorizer.IsVisible(chain))
+            {
+                throw new BadRequestException($"Filter parsing failed for ({key}, {value}).", new ForbiddenException("Selected property is forbidden."));
+            }
+
             var parts = (value ?? "").Split(',').ToList();
             var operand = ExtractFilterOperand(parts);
-
-            var chain = _expressionParser.ParseChain<TEntity>(key);
             var values = ConvertFilterValues(operand, chain, parts);
 
             return new WebFilter<TEntity>(chain, operand, values);
         }
 
-        public Filter<TEntity> Parse<TEntity>(HttpRequest request, IWebFilterConverter<TEntity> webFilterConverter)
-            where TEntity : class
-            => Parse(request, null, webFilterConverter);
+        public Filter<TEntity> Parse(HttpRequest request) => Parse(request, null);
 
-        public Filter<TEntity> Parse<TEntity>(HttpRequest request, ActionDescriptor action, IWebFilterConverter<TEntity> webFilterConverter)
-            where TEntity : class
+        public Filter<TEntity> Parse(HttpRequest request, ActionDescriptor action)
         {
             var keyValuePairs = request.Query.Where(kv => !string.IsNullOrWhiteSpace(kv.Key) && !Reserved.IsKeyword(kv.Key));
 
@@ -74,8 +81,8 @@ namespace Rdd.Web.Querying
                 keyValuePairs = keyValuePairs.Where(kvp => !queryActionParameters.Contains(kvp.Key));
             }
 
-            var filters = keyValuePairs.Select(kv => Parse<TEntity>(kv.Key, kv.Value));
-            return webFilterConverter.ToExpression(filters);
+            var filters = keyValuePairs.Select(kv => Parse(kv.Key, kv.Value));
+            return _webFilterConverter.ToExpression(filters);
         }
 
         protected virtual WebFilterOperand ExtractFilterOperand(List<string> parts)